+++ app/page.js
@@ -1,0 +1,28 @@
+import Cta from "@layouts/components/Cta";
+import Features from "@layouts/partials/Features";
+import HomeBanner from "@layouts/partials/HomeBanner";
+import SeoMeta from "@layouts/partials/SeoMeta";
+import ShortIntro from "@layouts/partials/ShortIntro";
+import SpecialFeatures from "@layouts/partials/SpecialFeatures";
+import Testimonial from "@layouts/partials/Testimonial";
+import { getListPage } from "@lib/contentParser";
+
+const Home = async () => {
+  const homepage = await getListPage("content/_index.md");
+  const { frontmatter } = homepage;
+  const { banner, brands, features, intro, speciality, testimonial } =
+    frontmatter;
+  return (
+    <>
+      <SeoMeta title="Home" />
+      <HomeBanner banner={banner} brands={brands} />
+      <Features features={features} />
+      <ShortIntro intro={intro} />
+      <SpecialFeatures speciality={speciality} />
+      <Testimonial testimonial={testimonial} />
+      <Cta />
+    </>
+  );
+};
+
+export default Home;
