+++ package.json
@@ -17,10 +17,10 @@
     "gray-matter": "^4.0.3",
     "gsap": "^3.11.4",
     "marked": "^4.2.12",
-    "next": "13.1.6",
+    "next": "^14.2.6",
     "next-mdx-remote": "^4.3.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-gtm-module": "^2.0.11",
     "react-icons": "^4.7.1",
     "react-lite-youtube-embed": "^2.3.52",
@@ -46,6 +46,6 @@
     "sass": "^1.57.1",
     "sharp": "^0.31.3",
     "tailwind-bootstrap-grid": "^5.0.1",
-    "tailwindcss": "^3.2.4"
+    "tailwindcss": "^3.4.10"
   }
 }
